+++ python/lsst/ts/MTDomeTrajectory/config_schema.py
@@ -1,5 +1,32 @@
+# This file is part of ts_MTDomeTrajectory.
+#
+# Developed for Vera C. Rubin Observatory Telescope and Site Systems.
+# This product includes software developed by the LSST Project
+# (https://www.lsst.org).
+# See the COPYRIGHT file at the top-level directory of this distribution
+# for details of code ownership.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+__all__ = ["CONFIG_SCHEMA"]
+
+import yaml
+
+CONFIG_SCHEMA = yaml.safe_load(
+    """
 $schema: http://json-schema.org/draft-07/schema#
-$id: https://github.com/lsst-ts/ts_MTDomeTrajectory/blob/master/schema/MTDomeTrajectory.yaml
+$id: https://github.com/lsst-ts/ts_MTDomeTrajectory/blob/master/python/lsst/ts/MTDomeTrajectory/config_schema.py  # noqa
 # title must end with one or more spaces followed by the schema version, which must begin with "v"
 title: MTDomeTrajectory v1
 description: Schema for MTDomeTrajectory configuration files
@@ -43,4 +70,5 @@ allOf:
           max_delta_elevation: 6
         additionalProperties: false
 additionalProperties: false
-
+"""
+)