+++ tests/test_csc.py
@@ -1,6 +1,6 @@
 # This file is part of ts_MTDomeTrajectory.
 #
-# Developed for the LSST Telescope and Site Systems.
+# Developed for Vera C. Rubin Observatory Telescope and Site Systems.
 # This product includes software developed by the LSST Project
 # (https://www.lsst.org).
 # See the COPYRIGHT file at the top-level directory of this distribution
@@ -155,10 +155,7 @@ class MTDomeTrajectoryTestCase(salobj.BaseCscTestCase, asynctest.TestCase):
             initial_state=salobj.State.STANDBY, config_dir=TEST_CONFIG_DIR
         ):
             self.assertEqual(self.csc.summary_state, salobj.State.STANDBY)
-            state = await self.remote.evt_summaryState.next(
-                flush=False, timeout=STD_TIMEOUT
-            )
-            self.assertEqual(state.summaryState, salobj.State.STANDBY)
+            await self.assert_next_summary_state(salobj.State.STANDBY)
 
             for bad_config_name in (
                 "no_such_file.yaml",
@@ -174,10 +171,7 @@ class MTDomeTrajectoryTestCase(salobj.BaseCscTestCase, asynctest.TestCase):
             self.remote.cmd_start.set(settingsToApply="valid.yaml")
             await self.remote.cmd_start.start(timeout=STD_TIMEOUT)
             self.assertEqual(self.csc.summary_state, salobj.State.DISABLED)
-            state = await self.remote.evt_summaryState.next(
-                flush=False, timeout=STD_TIMEOUT
-            )
-            self.assertEqual(state.summaryState, salobj.State.DISABLED)
+            await self.assert_next_summary_state(salobj.State.DISABLED)
             settings = await self.remote.evt_algorithm.next(
                 flush=False, timeout=STD_TIMEOUT
             )